Fix golang-project issue with the workspace change.
The new workspace function changes directory into the shadow workspace.
golang-project made a dockerfile in the workspace root, which was out of
scope for 'docker build'.

This change preserves the original workdir of the build step in $WORKSPACE,
so that golang-project can easily run 'docker build' from a context that
has all the files needed.

-------------
MOE_MIGRATED_REVID=149109941
APPROVED=