Adjusted clouduild.yaml to use `go test`.
The existing `bazel test` was causing problems for a challenging
test I added that needs to ensure that os.Exit(1) is invoked.
To perform this test, I used execute.Cmd() to launch another
process so that I could examine the exit code (and stderr). This
did not play nicely with Blaze's hermetic environments.