Fix ~/.bazelrc not being used in Bazel builder
By using `--announce_rc` we found that the `.bazelrc` added in this Dockerfile was not being used when running in cloud build.
This results in steps not sharing the output base, i.e. being slow.

Under Cloud Build the home folder is `/builder/home` whereas under Docker it's `/root`. Therefore the `.bazelrc` is not found.