Update Alternative Images section (#886)
Getting some feedback from a CE and customer that this section as phrased was confusing, and kind of suggested that the alternative images were also somehow 'official' and therefore supported by Google. Made a very small change to the header, and also removing the word 'official' from before 'community builders', in an attempt to avoid confusion in the short term. And we'll need to revisit all of this positioning as we fold our builders into the new Catalog strategy.