Update go version and dependencies for a few binaries.
The go version used for building the binaries as well as a number
of the dependencies are stale. This change updates the go version
used for building the binary to go1.21.x and also updates deps and
vendor directories using `go get -u ./...` `go mod tidy` and
`go mod vendor`. Most of the files in this change are from vendored
dependencies after the update.

Confirmed that after buildling these binaries, reported fixable CVEs
drops to 0 (gke-deploy, gcs-fetcher and gcs-uploader).