Enable fetching image from image/artifact registry
Prior to this, when we fetched the image to extract its digest,
we used the `auth.DefaultKeyChain` which looks at the docker config.
While this is fine when the gcloud sdk is installed in the environment,
it does not work without the use of gcloud in the environment and docker config.
We therefore need to try fetching it using the go-client library if the
default mechanism fails.